reconstruct headers
